chatbot-container/pyproject.toml

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"



[tool.poetry]
name = "chatbot"
version = "0.3.0"
description = "Chat Bot for Kubernetes"
authors = ["Ben Greene <BenJGreene@gmail.com>"]
readme = "README.md"
license = "GPL3"
repository = "https://github.com/PolecatWorks/chatbot"
include = [
    "LICENSE",
]

[tool.poetry.dependencies]
python = "^3.12"
click = "^8"
# "ruamel.yaml" = "^0.18.6"
pydantic-yaml = {extras = ["pyyaml"], version = "^1.4"}
aiohttp = "^3"
pydantic-settings = {extras = ["yaml"], version = "^2.6.1"}
pydantic-file-secrets = "^0.4.4"
prometheus-async = "^25.1.0"
botbuilder-integration-aiohttp= ">=4.14.0"
langchain-openai = "^0.3.18"
langchain-google-genai = "^2.1.5"
langchain = "^0.3.25"
langchain-mcp-adapters = "^0.1.7"


[tool.poetry.group.dev.dependencies]
aiohttp-devtools = "^1"
pytest = "^8"
Sphinx = "^8"
sphinx-autodoc-typehints = "^3.2.0"
pytest-sphinx = "^0.6"
flake8 = "^7.1.1"
flake8-black = "^0.3.6"
pylint = "^3.3.1"
sphinx-click = "^6.0.0"
pytest-click = "^1.1.0"
pytest-cov = "^6.1.1"
pytest-aiohttp = "^1"
pytest-watch = "^4.2.0"



[tool.poetry.scripts]
chatbot = "chatbot.cli:cli"
